portal: deploy new dashboard
As per usual nixos threatens my free time with promises of flawless
deployments.  Everyone has their heads in the clouds.  Not impressed.  I don't
appreciate buildNpmPackage complaining to me that it can't pull down a
dependency referenced in package.json and can't cache it either.  Isn't that
that the entire job of that function???  Good news, I didn't fall for the bait.
Instead, just rsynced over the output and called it a day.  Checkmate nixers.